feat: change int to short from UserInfo

+++ Code/EncodableData/UserInfo.cs
@@ -12,10 +12,10 @@ public class UserInfo : IEncodable
 	public ChatModeratorLevel? ChatModeratorLevel { get; set; }
 
 	[Encode(1)]
-	public int Deaths { get; set; }
+	public short Deaths { get; set; }
 
 	[Encode(2)]
-	public int Kills { get; set; }
+	public short Kills { get; set; }
 
 	[Encode(3)]
 	public byte Rank { get; set; }
